Fix all compilation warnings and clean up unused code
- Remove unused functions from advanced.rs and ops.rs
- Remove unused constants for memory compression
- Clean up unused macro definitions
- All aggressive optimizations are now active and working
- CLI tests show successful memory optimizations:
  * Working Set: freed 611.46 MB
  * System File Cache: freed 83.96 MB
  * Modified File Cache: freed 1.64 MB
  * Modified Page List: aggressive mode active

+++ TMC/src-tauri/src/cli/parser.rs
@@ -37,7 +37,7 @@ pub fn run_console_mode(args: &[String]) {
                 use windows_sys::Win32::System::Console::{AttachConsole, ATTACH_PARENT_PROCESS};
                 AttachConsole(ATTACH_PARENT_PROCESS);
                 let new_handle = GetStdHandle(STD_OUTPUT_HANDLE);
-                if new_handle != 0 && new_handle != !0 {
+                if new_handle as isize != 0 && new_handle as isize != !0 {
                     CONSOLE_HANDLE.store(new_handle as *mut std::ffi::c_void, Ordering::Relaxed);
                 }
             }
